+++ src/cascadia/WindowsTerminal/NonClientIslandWindow.cpp
@@ -80,7 +80,7 @@ void NonClientIslandWindow::MakeWindow() noexcept
                                          0,
                                          0,
                                          0,
-                                         GetWindowHandle(),
+                                         GetHandle(),
                                          nullptr,
                                          wil::GetModuleInstanceHandle(),
                                          this));
@@ -122,7 +122,7 @@ LRESULT NonClientIslandWindow::_InputSinkMessageHandler(UINT const message, WPAR
         POINT screenPt{ clientPt };
         if (ClientToScreen(_dragBarWindow.get(), &screenPt))
         {
-            auto parentWindow{ GetWindowHandle() };
+            auto parentWindow{ GetHandle() };
 
             // Hit test the parent window at the screen coordinates the user clicked in the drag input sink window,
             // then pass that click through as an NC click in that location.
@@ -569,7 +569,7 @@ int NonClientIslandWindow::_GetResizeHandleHeight() const noexcept
         // with that message at the time it was sent to handle the message
         // correctly.
         const auto screenPtLparam{ GetMessagePos() };
-        const LRESULT hitTest{ SendMessage(GetWindowHandle(), WM_NCHITTEST, 0, screenPtLparam) };
+        const LRESULT hitTest{ SendMessage(GetHandle(), WM_NCHITTEST, 0, screenPtLparam) };
         if (hitTest == HTTOP)
         {
             // We have to set the vertical resize cursor manually on
@@ -592,7 +592,7 @@ int NonClientIslandWindow::_GetResizeHandleHeight() const noexcept
         }
     }
 
-    return DefWindowProc(GetWindowHandle(), WM_SETCURSOR, wParam, lParam);
+    return DefWindowProc(GetHandle(), WM_SETCURSOR, wParam, lParam);
 }
 
 // Method Description:
