+++ src/interactivity/inc/IConsoleWindow.hpp
@@ -15,14 +15,12 @@ Author(s):
 
 #pragma once
 
-#include "IUiaWindow.h"
-
 // copied typedef from uiautomationcore.h
 typedef int EVENTID;
 
 namespace Microsoft::Console::Types
 {
-    class IConsoleWindow : public IUiaWindow
+    class IConsoleWindow
     {
     public:
         virtual ~IConsoleWindow() = 0;
@@ -38,8 +36,12 @@ namespace Microsoft::Console::Types
 
         virtual void SetIsFullscreen(const bool fFullscreenEnabled) = 0;
 
+        virtual void ChangeViewport(const SMALL_RECT NewWindow) = 0;
+
         virtual void CaptureMouse() = 0;
         virtual BOOL ReleaseMouse() = 0;
+
+        virtual HWND GetWindowHandle() const = 0;
 
         // Pass null.
         virtual void SetOwner() = 0;
@@ -62,6 +64,10 @@ namespace Microsoft::Console::Types
                                       const WORD wAbsoluteChange) = 0;
         virtual void VerticalScroll(const WORD wScrollCommand,
                                     const WORD wAbsoluteChange) = 0;
+
+        [[nodiscard]] virtual HRESULT SignalUia(_In_ EVENTID id) = 0;
+        [[nodiscard]] virtual HRESULT UiaSetTextAreaFocus() = 0;
+        virtual RECT GetWindowRect() const noexcept = 0;
     };
 
     inline IConsoleWindow::~IConsoleWindow() {}
