+++ src/types/sources.inc
@@ -42,7 +42,6 @@ SOURCES= \
     ..\Utf16Parser.cpp \
     ..\utils.cpp \
     ..\ThemeUtils.cpp \
-    ..\WindowUiaProviderBase.cpp \
     ..\ScreenInfoUiaProviderBase.cpp \
     ..\UiaTextRangeBase.cpp \
     ..\UiaTracing.cpp \
